all: add proxy and generic settings support (#2)
* seva-web : Add Support for Side-NavBar and Webproxy Page

Signed-off-by: Chirag Shilwant <c-shilwant@ti.com>

* Add a TexFormField, Save Icon and Rename webproxy_page.dart to settings_page.dart

Signed-off-by: Chirag Shilwant <c-shilwant@ti.com>

* Add Support for Webproxy via CLI and Flutter Page

Signed-off-by: Chirag Shilwant <c-shilwant@ti.com>

* seva-launcher: Rework the proxy settings

Use generic proxy function and cleanup some of the logic. Adds a
ProxySettings structure to contain the settings in an somewhat agnostic
form.

Signed-off-by: Randolph Sapp <rs@ti.com>

* seva-web: Rename navigation menu

Rename the NavDrawer to NavigationMenu and rename the src files
accordingly.

Signed-off-by: Randolph Sapp <rs@ti.com>

* seva-web: Remove theme override on settings

Also make settings variables and functions more generic.

Signed-off-by: Randolph Sapp <rs@ti.com>

* seva-web: More proxy cleanup

Remove old deprecated page.

Rename variables to use the following notation:
  section_subsection_subsubsection

Signed-off-by: Randolph Sapp <rs@ti.com>

* seva-web: Fix the proxy validator

The proxy validator logic was wrong. This inverts the logic and attempts
to simplify it.

Signed-off-by: Randolph Sapp <rs@ti.com>

* seva-web: Make the settings more robust

Make the settings page build itself out of settings subsections and move
some widgets around to fit the feel of the rest of the app a little
more.

Signed-off-by: Randolph Sapp <rs@ti.com>

---------

Signed-off-by: Chirag Shilwant <c-shilwant@ti.com>
Signed-off-by: Randolph Sapp <rs@ti.com>
Co-authored-by: Chirag Shilwant <c-shilwant@ti.com>
Co-authored-by: Randolph Sapp <rs@ti.com>